+++ app/src/main/java/elder/ly/mobile/MainActivity.kt
@@ -9,12 +9,10 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -41,9 +39,52 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 fun App() {
-    Text(
-        text = "Elderly"
-    )
+    Search()
+}
+
+@Composable
+fun Search(){
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .shadow(
+                elevation = 4.dp,
+                shape = RoundedCornerShape(16.dp),
+                clip = false
+            )
+            .background(
+                color = Color.White,
+                shape = RoundedCornerShape(8.dp)
+            )
+            .border(
+                width = 1.dp,
+                color = tertiaryContainerLight,
+                shape = RoundedCornerShape(4.dp)
+            )
+            .padding(8.dp)
+    ) {
+        Image(
+            painter = painterResource(id = ic_pesquisar),
+            contentDescription = "Ícone de pesquisa",
+            modifier = Modifier
+                .size(48.dp)
+                .align(alignment = Alignment.CenterStart)
+                .padding(start = 8.dp)
+        )
+        Column(
+            modifier = Modifier
+                .align(Alignment.Center)
+                .padding(start = 32.dp)
+        ) {
+            Text(
+                text = "20/08 9:00 - 20/08 20:00",
+                fontWeight = FontWeight.Bold
+            )
+            Text(
+                text = "Troca de Fralda, Medicação, Bingo"
+            )
+        }
+    }
 }
 
 @Preview(showBackground = true)
